requirements.txt
streamlit==1.56.0
tensorflow==2.16.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
